WIP: refactoring game/test for readability

+++ services/backend/build/init.sh
@@ -4,6 +4,10 @@ set -eu
 
 mkdir -p /backend/DB/website
 
+if [ "$APP_MODE" = "test" ]; then
+    echo "Running Tests..."
+    exec conda run --no-capture-output -n backend python /backend/manage.py test game
+    exit 0
 rm -f /backend/DB/website/db.sqlite3
 find /backend -path "*/migrations/0*" -delete
 
@@ -11,10 +15,6 @@ conda run -n backend python /backend/manage.py makemigrations
 conda run -n backend python /backend/manage.py migrate
 conda run -n backend python /backend/manage.py collectstatic --noinput
 
-if [ "$APP_MODE" = "test" ]; then
-    echo "Running Tests..."
-    exec conda run --no-capture-output -n backend python /backend/manage.py test chat.tests.ChatViewsTests.test_room_not_found_returns_404
-    exit 0
 fi
 
 conda run -n backend bash -c "python /backend/manage.py shell < /backend/seed.py"
